blueprint_dependencies
